fix(user): fixed the user saved data bug

+++ src/app/pages/user/user-workout/user-workout.page.ts
@@ -6,6 +6,8 @@ import {WorkoutInterface} from '../../../interfaces/workout.interface';
 import {NavController} from '@ionic/angular';
 import {SpinnerService} from '../../../services/spinner.service';
 import {ToastService} from '../../../services/toast.service';
+import {UserInterface} from '../../../interfaces/user.interface';
+import {AuthService} from '../../../services/auth.service';
 
 
 @Component({
@@ -17,6 +19,7 @@ export class UserWorkoutPage {
   workoutId: string | null = null;
   userWorkout: UserWorkoutInterface | null = null;
   workouts: WorkoutInterface[] = [];
+  user: UserInterface | null = null;
 
 
   constructor(
@@ -24,24 +27,30 @@ export class UserWorkoutPage {
     private route: ActivatedRoute,
     private navControl: NavController,
     private spinnerService: SpinnerService,
-    private toastService: ToastService
+    private toastService: ToastService,
+    private authService: AuthService
   ) { }
 
 
   ionViewWillEnter(): void {
     this.workoutId = this.route.snapshot.paramMap.get('id');
+    this.authService.user.subscribe((user: UserInterface | null): void => {
+      this.user = user;
+    });
     this.getUserWorkoutDoc();
   }
 
 
   getUserWorkoutDoc(): void {
     if (this.workoutId) {
-      this.userWorkoutService.getUserWorkoutById(this.workoutId).then((workout: UserWorkoutInterface): void => {
-        this.userWorkout = workout;
-        this.workouts = workout.workouts as unknown as WorkoutInterface[];
-      }).catch(err => {
-        console.error('Error in getUserWorkoutDoc: ', err);
-      });
+      if (this.user && this.user.uid) {
+        this.userWorkoutService.getUserWorkoutById(this.workoutId, this.user.uid).then((workout: UserWorkoutInterface): void => {
+          this.userWorkout = workout;
+          this.workouts = workout.workouts as unknown as WorkoutInterface[];
+        }).catch(err => {
+          console.error('Error in getUserWorkoutDoc: ', err);
+        });
+      }
     }
   }
 
@@ -57,34 +66,38 @@ export class UserWorkoutPage {
 
   deleteWorkout(workoutToRemoveId: string): void {
     if (this.userWorkout && this.userWorkout.id) {
-      this.spinnerService.show('Excluindo...');
-      const workoutId: string = this.userWorkout.id;
-      this.userWorkoutService.deleteWorkoutFromArray(workoutId, workoutToRemoveId).then((): void => {
-        this.workouts = [];
-        this.userWorkout = null;
-        this.getUserWorkoutDoc();
-        this.spinnerService.hide();
-        void this.toastService.presentSuccessToast('Treino excluido com sucesso!');
-      }).catch(err => {
-        console.error('Error in deleteWorkout', err);
-        this.spinnerService.hide();
-        void this.toastService.presentErrorToast('Erro ao excluir o treino!');
-      });
+      if (this.user && this.user.uid) {
+        this.spinnerService.show('Excluindo...');
+        const workoutId: string = this.userWorkout.id;
+        this.userWorkoutService.deleteWorkoutFromArray(workoutId, workoutToRemoveId, this.user.uid).then((): void => {
+          this.workouts = [];
+          this.userWorkout = null;
+          this.getUserWorkoutDoc();
+          this.spinnerService.hide();
+          void this.toastService.presentSuccessToast('Treino excluido com sucesso!');
+        }).catch(err => {
+          console.error('Error in deleteWorkout', err);
+          this.spinnerService.hide();
+          void this.toastService.presentErrorToast('Erro ao excluir o treino!');
+        });
+      }
     }
   }
 
 
   deleteAllWorkout(): void {
     if (this.userWorkout && this.userWorkout.id) {
-      this.spinnerService.show('Excluindo todo o treino...');
-      this.userWorkoutService.deleteWorkout(this.userWorkout.id).then((): void => {
-        void this.toastService.presentSuccessToast('Todo o treino foi excluido!');
-        this.navControl.navigateForward('/home').then(() => this.spinnerService.hide());
-      }).catch(err => {
-        this.spinnerService.hide();
-        console.error('Error in deleteWorkout', err);
-        void this.toastService.presentErrorToast('Erro ao excluir o treino todo!');
-      });
+      if (this.user && this.user.uid) {
+        this.spinnerService.show('Excluindo todo o treino...');
+        this.userWorkoutService.deleteWorkout(this.userWorkout.id, this.user.uid).then((): void => {
+          void this.toastService.presentSuccessToast('Todo o treino foi excluido!');
+          this.navControl.navigateForward('/home').then(() => this.spinnerService.hide());
+        }).catch(err => {
+          this.spinnerService.hide();
+          console.error('Error in deleteWorkout', err);
+          void this.toastService.presentErrorToast('Erro ao excluir o treino todo!');
+        });
+      }
     }
   }
